refactor: moved read state_sync into check_tx
